Day 12 - Part 2
So for Part 1 I wrote a parser that looks through the txt file. For Part
2 I actually had JS take care of the .json file for me which I then
iterated through to find the numbers that satisfy the needed condition.
Also as a side note: I have no idea why on line 13 of "Part 2.js" the
triple equality works, but double will not. I only point this out since
it was the thing I was stuck on for a little too long lol.